[test - #84] api 통합으로 인한 서비스 테스트 추가

+++ src/test/java/com/dateplan/dateplan/service/schedule/ScheduleServiceTest.java
@@ -18,7 +18,6 @@ import com.dateplan.dateplan.global.constant.Gender;
 import com.dateplan.dateplan.global.constant.Operation;
 import com.dateplan.dateplan.global.constant.RepeatRule;
 import com.dateplan.dateplan.global.constant.Resource;
-import com.dateplan.dateplan.global.exception.ErrorCode.DetailMessage;
 import com.dateplan.dateplan.global.exception.auth.NoPermissionException;
 import com.dateplan.dateplan.global.exception.schedule.ScheduleNotFoundException;
 import com.dateplan.dateplan.global.util.ScheduleDateUtil;
@@ -32,7 +31,9 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -59,9 +60,8 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 
 		@BeforeEach
 		void setUp() {
-			member = createMember("nickname");
+			member = memberRepository.save(createMember("nickname"));
 			MemberThreadLocal.set(member);
-			memberRepository.save(member);
 		}
 
 		@AfterEach
@@ -141,41 +141,57 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 	@Nested
 	class UpdateSchedule {
 
+		private static final String NEED_SCHEDULE = "needSchedule";
+
 		Member member;
-		Schedule schedule;
+		List<Schedule> schedules;
 
 		@BeforeEach
-		void setUp() {
+		void setUp(TestInfo testInfo) {
 			member = memberRepository.save(createMember("nickname"));
 			MemberThreadLocal.set(member);
 
-			SchedulePattern schedulePattern = schedulePatternRepository.save(
-				SchedulePattern.builder()
-					.member(member)
-					.repeatRule(RepeatRule.N)
-					.repeatStartDate(LocalDate.now())
-					.repeatEndDate(DateConstants.CALENDER_END_DATE)
-					.build()
-			);
-			schedule = scheduleRepository.save(createSchedule(schedulePattern));
+			if (testInfo.getTags().contains(NEED_SCHEDULE)) {
+				SchedulePattern schedulePattern = schedulePatternRepository.save(
+					SchedulePattern.builder()
+						.member(member)
+						.repeatRule(RepeatRule.N)
+						.repeatStartDate(LocalDate.now())
+						.repeatEndDate(DateConstants.CALENDER_END_DATE)
+						.build()
+				);
+				schedules = List.of(
+					createSchedule(schedulePattern),
+					createSchedule(schedulePattern),
+					createSchedule(schedulePattern),
+					createSchedule(schedulePattern),
+					createSchedule(schedulePattern)
+				);
+				scheduleRepository.saveAll(schedules);
+			}
 		}
 
 		@AfterEach
-		void tearDown() {
+		void tearDown(TestInfo testInfo) {
+			if (testInfo.getTags().contains(NEED_SCHEDULE)) {
+				scheduleRepository.deleteAllInBatch();
+				schedulePatternRepository.deleteAllInBatch();
+			}
 			MemberThreadLocal.remove();
-			scheduleRepository.deleteAllInBatch();
-			schedulePatternRepository.deleteAllInBatch();
 			memberRepository.deleteAllInBatch();
 		}
 
-		@DisplayName("요청 시, 요청의 내용으로 일정이 수정된다.")
+		@Tag(NEED_SCHEDULE)
+		@DisplayName("단일 일정 수정하는 요청 시, 요청의 내용으로 해당 일정이 수정된다.")
 		@Test
-		void successWithValidRequest() {
+		void successWithSingleUpdateRequest() {
 			// Given
 			ScheduleUpdateServiceRequest request = createScheduleUpdateServiceRequest();
+			Schedule schedule = schedules.get(0);
 
 			// When
-			scheduleService.updateSchedule(member.getId(), schedule.getId(), request, member);
+			scheduleService.updateSchedule(member.getId(), schedule.getId(), request, member,
+				false);
 
 			// Then
 			Schedule updatedSchedule = scheduleRepository.findById(schedule.getId()).get();
@@ -185,6 +201,38 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 			assertThat(updatedSchedule.getLocation()).isEqualTo(request.getLocation());
 			assertThat(updatedSchedule.getStartDateTime()).isEqualTo(request.getStartDateTime());
 			assertThat(updatedSchedule.getEndDateTime()).isEqualTo(request.getEndDateTime());
+		}
+
+		@Tag(NEED_SCHEDULE)
+		@DisplayName("반복 일정 수정하는 요청 시, 요청의 내용으로 모든 일정이 수정된다.")
+		@Test
+		void successWithRepeatUpdateRequest() {
+			// Given
+			ScheduleUpdateServiceRequest request = createScheduleUpdateServiceRequest();
+			Schedule schedule = schedules.get(0);
+
+			// When
+			scheduleService.updateSchedule(member.getId(), schedule.getId(), request, member,
+				true);
+
+			// Then
+			List<Schedule> updatedSchedules = scheduleRepository.findBySchedulePatternId(
+				schedule.getSchedulePattern().getId());
+
+			long startTimeDiff = ChronoUnit.MINUTES.between(schedule.getStartDateTime(),
+				request.getStartDateTime());
+			long endTimeDiff = ChronoUnit.MINUTES.between(schedule.getEndDateTime(),
+				request.getEndDateTime());
+
+			for (int i = 0; i < updatedSchedules.size(); i++) {
+				assertThat(updatedSchedules.get(i).getTitle()).isEqualTo(request.getTitle());
+				assertThat(updatedSchedules.get(i).getContent()).isEqualTo(request.getContent());
+				assertThat(updatedSchedules.get(i).getLocation()).isEqualTo(request.getLocation());
+				assertThat(updatedSchedules.get(i).getStartDateTime()).isEqualTo(
+					schedules.get(i).getStartDateTime().plus(startTimeDiff, ChronoUnit.MINUTES));
+				assertThat(updatedSchedules.get(i).getEndDateTime()).isEqualTo(
+					schedules.get(i).getEndDateTime().plus(endTimeDiff, ChronoUnit.MINUTES));
+			}
 		}
 
 		@DisplayName("현재 로그인한 회원의 id와 요청의 member_id가 다르면 실패한다.")
@@ -198,8 +246,7 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 			NoPermissionException exception = new NoPermissionException(Resource.MEMBER,
 				Operation.UPDATE);
 			assertThatThrownBy(() ->
-				scheduleService.updateSchedule(member.getId() + 100, schedule.getId(), request,
-					member))
+				scheduleService.updateSchedule(member.getId() + 100, 1L, request, member, true))
 				.isInstanceOf(exception.getClass())
 				.hasMessage(exception.getMessage());
 
@@ -215,7 +262,7 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 			ScheduleNotFoundException exception = new ScheduleNotFoundException();
 			assertThatThrownBy(() ->
 				scheduleService.updateSchedule(
-					member.getId(), schedule.getId() + 100, request, member))
+					member.getId(), 100000L, request, member, true))
 				.isInstanceOf(exception.getClass())
 				.hasMessage(exception.getMessage());
 
@@ -228,8 +275,8 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 			.title("title")
 			.content("content")
 			.location("location")
-			.startDateTime(LocalDateTime.now())
-			.endDateTime(LocalDateTime.now().plusDays(5))
+			.startDateTime(LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS))
+			.endDateTime(LocalDateTime.now().plusDays(5).truncatedTo(ChronoUnit.SECONDS))
 			.build();
 	}
 
@@ -258,8 +305,8 @@ public class ScheduleServiceTest extends ServiceTestSupport {
 	private ScheduleServiceRequest createScheduleServiceRequest(RepeatRule repeatRule) {
 		return ScheduleServiceRequest.builder()
 			.title("title")
-			.startDateTime(LocalDateTime.now().with(LocalTime.MIN))
-			.endDateTime(LocalDateTime.now().with(LocalTime.MAX))
+			.startDateTime(LocalDateTime.now().with(LocalTime.MIN).truncatedTo(ChronoUnit.SECONDS))
+			.endDateTime(LocalDateTime.now().with(LocalTime.MAX).truncatedTo(ChronoUnit.SECONDS))
 			.repeatRule(repeatRule)
 			.build();
 	}
